Updating dataset load
an update that changes the way in which the dataset is loaded. the dataset is no longer loaded at once, but the image paths to be used are saved and the images are loaded each time when they are needed (pattern training). in this way the programme increases its efficiency for VRAM at the expense of training time.